cleanup(angular): host generator
