+++ packages/angular/src/generators/host/host.ts
@@ -1,18 +1,8 @@
-import {
-  formatFiles,
-  generateFiles,
-  getProjects,
-  joinPathFragments,
-  names,
-  readProjectConfiguration,
-  Tree,
-} from '@nrwl/devkit';
+import { formatFiles, getProjects, Tree } from '@nrwl/devkit';
 import type { Schema } from './schema';
 import applicationGenerator from '../application/application';
 import remoteGenerator from '../remote/remote';
 import { normalizeProjectName } from '../utils/project';
-import * as ts from 'typescript';
-import { addRoute } from '../../utils/nx-devkit/route-utils';
 import { setupMf } from '../setup-mf/setup-mf';
 import { E2eTestRunner } from '../../utils/test-runners';
 
@@ -66,81 +56,11 @@ export async function host(tree: Tree, options: Schema) {
     });
   }
 
-  routeToNxWelcome(tree, options);
-
   if (!options.skipFormat) {
     await formatFiles(tree);
   }
 
   return installTask;
-}
-
-function routeToNxWelcome(tree: Tree, options: Schema) {
-  const { sourceRoot } = readProjectConfiguration(
-    tree,
-    normalizeProjectName(options.name, options.directory)
-  );
-
-  const remoteRoutes =
-    options.remotes && Array.isArray(options.remotes)
-      ? options.remotes.reduce(
-          (routes, remote) =>
-            `${routes}\n<li><a routerLink='${normalizeProjectName(
-              remote,
-              options.directory
-            )}'>${names(remote).className}</a></li>`,
-          ''
-        )
-      : '';
-
-  tree.write(
-    joinPathFragments(sourceRoot, 'app/app.component.html'),
-    `<ul class="remote-menu">
-<li><a routerLink='/'>Home</a></li>
-${remoteRoutes}
-</ul>
-<router-outlet></router-outlet>
-`
-  );
-
-  const pathToHostRootRoutingFile = joinPathFragments(
-    sourceRoot,
-    'app/app.routes.ts'
-  );
-
-  const hostRootRoutingFile = tree.read(pathToHostRootRoutingFile, 'utf-8');
-
-  let sourceFile = ts.createSourceFile(
-    pathToHostRootRoutingFile,
-    hostRootRoutingFile,
-    ts.ScriptTarget.Latest,
-    true
-  );
-
-  addRoute(
-    tree,
-    pathToHostRootRoutingFile,
-    `{
-      path: '',
-      component: NxWelcomeComponent
-    }`
-  );
-
-  tree.write(
-    pathToHostRootRoutingFile,
-    `import { NxWelcomeComponent } from './nx-welcome.component';
-    ${tree.read(pathToHostRootRoutingFile, 'utf-8')}`
-  );
-
-  generateFiles(
-    tree,
-    joinPathFragments(__dirname, 'files'),
-    joinPathFragments(sourceRoot, 'app'),
-    {
-      appName: normalizeProjectName(options.name, options.directory),
-      tmpl: '',
-    }
-  );
 }
 
 export default host;
