+++ packages/angular/src/generators/setup-mf/files/host-files/app.component.spec.ts__tmpl__
