+++ packages/angular/src/generators/setup-mf/lib/index.ts
@@ -8,3 +8,4 @@ export * from './get-remotes-with-ports';
 export * from './set-tsconfig-target';
 export * from './setup-host-if-dynamic';
 export * from './setup-serve-target';
+export * from './update-host-app-routes';
